+++ sandbox/pages/configurator.tsx
@@ -1,11 +1,12 @@
 /* eslint-disable @next/next/no-img-element */
 
-import { Heading } from '@chakra-ui/react';
+import { Heading, Link } from '@chakra-ui/react';
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import React from 'react';
 import CanvasTitle from '../components/CanvasTitle';
 import CodeTitle from '../components/CodeTitle';
+import { Menu } from '../components/core/Menu';
 import DefaultVariablesValuesInput from '../components/DefaultVariablesValuesInput';
 import DefaultVariablesValuesRenderer from '../components/DefaultVariablesValuesRenderer';
 import InsertionModeInput from '../components/InsertionModeInput';
@@ -57,6 +58,7 @@ const Configurator: NextPage = () => {
               <img src="/logo.png" alt="logo" className={styles.logo} /> Git
               commit message configurator
             </Heading>
+            <Menu />
             <p className={styles.presentation}>
               This website as goal to generate settings for{' '}
               <a href="https://marketplace.visualstudio.com/items?itemName=rioukkevin.vscode-git-commit">